Centralize CLI commands.
* Also centralize plugin loading.
