+++ src/modelplane/runways/__init__.py
@@ -1,0 +1,4 @@
+from modelgauge.load_plugins import load_plugins
+
+
+load_plugins(disable_progress_bar=True)
