+++ src/modelplane/runways/responder.py
@@ -3,10 +3,8 @@
 import pathlib
 import tempfile
 
-import click
 import mlflow
 
-from modelgauge.load_plugins import load_plugins
 from modelgauge.pipeline_runner import PromptRunner
 from modelgauge.sut_registry import SUTS
 
@@ -16,60 +14,6 @@ from modelplane.runways.utils import (
     is_debug_mode,
     setup_sut_credentials,
 )
-from modelplane.utils.env import load_from_dotenv
-
-load_plugins(disable_progress_bar=True)
-
-
-@click.command(name="get-sut-responses")
-@click.option(
-    "--sut_id",
-    type=str,
-    required=True,
-    help="The SUT UID to use.",
-)
-@click.option(
-    "--prompts",
-    type=str,
-    required=True,
-    help="The path to the input prompts file.",
-)
-@click.option(
-    "--experiment",
-    type=str,
-    required=True,
-    help="The experiment name to use. If the experiment does not exist, it will be created.",
-)
-@click.option(
-    "--cache_dir",
-    type=str,
-    default=None,
-    help="The cache directory. Defaults to None. Local directory used to cache LLM responses.",
-)
-@click.option(
-    "--n_jobs",
-    type=int,
-    default=1,
-    help="The number of jobs to run in parallel. Defaults to 1.",
-)
-@load_from_dotenv
-def get_sut_responses(
-    sut_id: str,
-    prompts: str,
-    experiment: str,
-    cache_dir: str | None = None,
-    n_jobs: int = 1,
-):
-    """
-    Run the pipeline to get responses from SUTs.
-    """
-    return respond(
-        sut_id=sut_id,
-        prompts=prompts,
-        experiment=experiment,
-        cache_dir=cache_dir,
-        n_jobs=n_jobs,
-    )
 
 
 def respond(
